requirements.txt
flask
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
werkzeug==2.3.7
PyPDF2==3.0.1
google-cloud-documentai==2.20.1
google-cloud-language==2.9.1